refactor: 수익금 타입 BigDecimal로 변경

+++ src/main/java/blackjack/domain/judgement/GameResult.java
@@ -1,19 +1,20 @@
 package blackjack.domain.judgement;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 
 public enum GameResult {
 
-    BLACKJACK("블랙잭", 1.5),
-    WIN("승", 1.0),
-    DRAW("무", 0),
-    LOSE("패", -1.0);
+    BLACKJACK("블랙잭", BigDecimal.valueOf(1.5)),
+    WIN("승", BigDecimal.valueOf(1.0)),
+    DRAW("무", BigDecimal.valueOf(0)),
+    LOSE("패", BigDecimal.valueOf(-1.0));
 
     private final String name;
-    private final double dividendRate;
+    private final BigDecimal dividendRate;
 
-    GameResult(String name, double dividendRate) {
+    GameResult(String name, BigDecimal dividendRate) {
         this.name = name;
         this.dividendRate = dividendRate;
     }
@@ -33,7 +34,7 @@ public enum GameResult {
             .orElse(DRAW);
     }
 
-    public double calculateProfit(BettingMoney bettingMoney) {
-        return bettingMoney.multiply(dividendRate);
+    public BigDecimal calculateProfit(BettingMoney bettingMoney) {
+        return bettingMoney.toBigDecimal().multiply(dividendRate);
     }
 }
